Verplaats Pion.all() naar Pionnen factory
