+++ src/main/java/ipsen1/quarto/factory/Pionnen.java
@@ -1,6 +1,7 @@
 package ipsen1.quarto.factory;
 
 import ipsen1.quarto.business.Pion;
+import static ipsen1.quarto.business.Pion.*;
 
 import java.util.ArrayList;
 
@@ -13,5 +14,27 @@ public class Pionnen extends ArrayList<Pion> {
     public Pion getPion() {
         // TODO: Implementeer mij
         return new Pion();
+    }
+
+    /**
+     * Return an array of all unique pawn combinations.
+     * @return Pion[]
+     */
+    public static Pion[] all() {
+        Pion[] pionnen = new Pion[16];
+        int i = 0;
+
+        for(Vorm v : Vorm.values()) {
+            for (Kleur k : Kleur.values()) {
+                for (Hoogte h : Hoogte.values()) {
+                    for (Hol ho : Hol.values()) {
+                        pionnen[i] = new Pion(v, k, h, ho);
+                        i++;
+                    }
+                }
+            }
+        }
+
+        return pionnen;
     }
 }
